Initial refactoring into two key modules
