+++ prediction_api/requirements.txt
@@ -3,11 +3,11 @@ Flask==2.2.2
 itsdangerous==2.1.2
 Jinja2==3.1.2
 joblib==1.2.0
-MarkupSafe==2.1.1
-mysql-connector-python==8.0.31
-numpy==1.23.5
-protobuf==3.20.1
+MarkupSafe==2.1.2
+mysql-connector-python==8.0.32
+numpy==1.24.1
+protobuf==3.20.3
 scikit-learn==1.2.0
-scipy==1.9.3
+scipy==1.10.0
 threadpoolctl==3.1.0
 Werkzeug==2.2.2
