+++ web_app/database/controllers/Location.controller.js
