+++ web_app/database/controllers/WeatherInterval.controller.js
