+++ web_app/database/models/WeatherInterval.model.js
